\name{IVGC-package}
\alias{IVGC-package}
\alias{IVGC}
\docType{package}
\title{
\packageTitle{IVGC}
}
\description{
\packageDescription{IVGC}
}
\details{
\tabular{ll}{
Package: \tab IVGC\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2019-03-13\cr
License: \tab GPL-2\cr
}
}
\references{
Gao, B., Liu, X., Li, H. and Cui, Y. (2019). Integrative Analysis of Genetical Genomics Data Incorporating Network Structures. Manuscript.
}

\keyword{ package }

